+++ src/constants/ActionTypes.js
@@ -1,4 +1,4 @@
 export const UPDATE_AUTH_USER = "update_auth_user";
 
-export const GET_BOARD = "get_board";
-export const ADD_BOARD = "add_board";
+export const GET_PRODUCT_LIST = "get_product_list";
+export const ADD_PRODUCT = "add_product";
