--- src/services/axios.js
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const instance = axios.create({
-  baseURL: "http://13.125.219.196",
-});
-
-export { instance };
