--- src/shared/Cookie.js
@@ -1,14 +1,0 @@
-const getCookie = (name) => {
-  let value = "; " + document.cookie;
-  let parts = value.split(`; ${name}=`);
-
-  if (parts.length === 2) {
-    return parts.pop().split(":").shift();
-  }
-};
-
-const setCookie = (name, value) => {
-  document.cookie = `${name}=${value};`;
-};
-
-export { getCookie, setCookie };
